Refactor complainant details configuration and validation logic
- Simplified the editComplainantDetailsConfig by utilizing shared form step functions for better maintainability.
- Enhanced editProfileValidationUtils to use new validation functions for checking mobile number duplicates and respondent inquiry affidavit validation.
- Updated ordersCreateConfig to leverage new utility functions for building order reference fields and hearing date fields, reducing code duplication.
- Introduced shared configuration for order reference fields and hearing date fields in ordersCreateConfigShared.js to streamline order form configurations.
