+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/pages/citizen/FileCase/Config/complaindetailsConfig.js
@@ -1,341 +1,25 @@
 import {
-  complainantStandardNameFields,
-  fileCaseAddressLineInputs,
-  fileCaseAddressMatchIdNote,
-  fileCaseCompanyAddressLineInputs,
-  fileCaseCurrentResidentialAddressInputs,
-  verificationDisableConfigFields,
+  buildComplainantCompanyDetailStep,
+  complainantEntityTypeFormStep,
+  complainantIdFormStep,
+  complainantVerificationFormStep,
+  complainantDesignationFormStep,
+  fileCaseComplainantAddressFormStep,
+  fileCaseComplainantCompanyAddressFormStep,
+  fileCaseComplainantNameAgeFormStep,
+  fileCaseComplainantTypeFormStep,
 } from "../../../../configs/shared/partyFormFieldsShared";
 
 const complainantDetailsFormConfig = [
-  {
-    body: [
-      {
-        key: "complainantType",
-        head: "SELECT_COMPLAINANT_TYPE",
-        name: "complainantType",
-        type: "component",
-        notes: {
-          key: "personalDetailsNote",
-          type: "component",
-          component: "SelectCustomNote",
-          populators: {
-            inputs: [
-              {
-                type: "InfoComponent",
-                infoText: "CS_PLEASE_CONTACT_NYAY_MITRA_TEXT",
-                infoHeader: "CS_PLEASE_COMMON_NOTE",
-                infoTooltipMessage: "CS_NOTE_TOOLTIP_RESPONDENT_PERSONAL_DETAILS",
-              },
-            ],
-          },
-          withoutLabel: true,
-        },
-        component: "CustomRadioInfoComponent",
-        populators: {
-          type: "radioButton",
-          error: "CORE_REQUIRED_FIELD_ERROR",
-          label: "SELECT_COMPLAINANT_TYPE",
-          required: false,
-          mdmsConfig: {
-            select: "(data) => {return data['case'].ComplainantRespondentType?.map((item) => {return item;});}",
-            masterName: "ComplainantRespondentType",
-            moduleName: "case",
-          },
-          optionsKey: "code",
-          customStyle: {
-            gap: "40px",
-            alignItems: "center",
-            flexDirection: "row",
-          },
-          isDependent: true,
-          isMandatory: true,
-        },
-        isMandatory: true,
-        withoutLabel: true,
-        resetFormData: true,
-        noteDependentOn: "complainantVerification.individualDetails",
-      },
-    ],
-  },
-  {
-    body: [
-      {
-        key: "complainantTypeOfEntity",
-        type: "dropdown",
-        label: "TYPE_OF_ENTITY",
-        populators: {
-          name: "complainantTypeOfEntity",
-          type: "radioButton",
-          error: "CORE_REQUIRED_FIELD_ERROR",
-          label: "SELECT_RESPONDENT_TYPE",
-          styles: {
-            maxWidth: "100%",
-            marginBottom: "10px",
-          },
-          required: false,
-          mdmsConfig: {
-            select: "(data) => {return data['case'].TypeOfEntity?.map((item) => {return item;});}",
-            masterName: "TypeOfEntity",
-            moduleName: "case",
-          },
-          optionsKey: "code",
-          isMandatory: true,
-        },
-        isMandatory: true,
-      },
-    ],
-    dependentKey: {
-      complainantType: ["showCompanyDetails"],
-    },
-  },
-  {
-    body: [
-      {
-        key: "complainantVerification",
-        name: "mobileNumber",
-        type: "component",
-        error: "ERR_HRMS_INVALID_MOB_NO",
-        label: "CS_COMPLAINANT_MOBILE_NUMBER",
-        component: "VerifyPhoneNumber",
-        populators: {},
-        validation: {
-          pattern: {
-            masterName: "commonUiConfig",
-            moduleName: "patternValidation",
-            patternType: "contact",
-          },
-          required: true,
-          maxLength: 10,
-          minLength: 10,
-        },
-        isMandatory: true,
-        updateLabel: {
-          key: "label",
-          value: "CS_REPRESENTATIVE_MOBILE_NUMBER",
-        },
-        defaultLabel: {
-          key: "label",
-          value: "CS_COMPLAINANT_MOBILE_NUMBER",
-        },
-        withoutLabel: true,
-        updateLabelOn: "complainantType.showCompanyDetails",
-        componentInFront: "+91",
-        disableConfigKey: "individualDetails",
-        disableConfigFields: verificationDisableConfigFields,
-        isVerifiedOtpDisabledKey: "isDuplicateNumber",
-      },
-    ],
-    dependentKey: {
-      complainantType: ["commonFields"],
-    },
-  },
-  {
-    body: [
-      {
-        key: "complainantId",
-        type: "component",
-        component: "VerificationComponent",
-        populators: {
-          name: "complainantId",
-          inputs: [
-            {
-              name: "complainantId",
-              label: "COMPLAINANT_ID",
-              updateLabel: {
-                key: "label",
-                value: "CS_ENTITY_ID",
-              },
-              defaultLabel: {
-                key: "label",
-                value: "COMPLAINANT_ID",
-              },
-              updateLabelOn: "complainantType.showCompanyDetails",
-              verificationOn: "complainantVerification.individualDetails",
-            },
-          ],
-          customStyle: {
-            marginTop: 20,
-          },
-        },
-        isMandatory: true,
-        withoutLabel: true,
-      },
-    ],
-    dependentKey: {
-      complainantType: ["commonFields"],
-    },
-  },
-  {
-    body: [
-      ...complainantStandardNameFields,
-      {
-        type: "text",
-        label: "AGE",
-        populators: {
-          name: "complainantAge",
-          error: "AGE_VALIDATION",
-          validation: {
-            maxLength: 3,
-            patternType: "Number",
-          },
-        },
-        isMandatory: true,
-      },
-    ],
-    head: "CS_COMMON_COMPLAINANT_DETAIL",
-    updateLabel: {
-      key: "head",
-      value: "CS_COMMON_ENTITY_DETAIL",
-    },
-    defaultLabel: {
-      key: "head",
-      value: "CS_COMMON_COMPLAINANT_DETAIL",
-    },
-    dependentKey: {
-      complainantType: ["commonFields"],
-    },
-    updateLabelOn: "complainantType.showCompanyDetails",
-  },
-  {
-    body: [
-      {
-        type: "text",
-        label: "DESIGNATION",
-        populators: {
-          name: "complainantDesignation",
-          error: "CORE_REQUIRED_FIELD_ERROR",
-          validation: {
-            title: "",
-            pattern: {
-              message: "CORE_COMMON_APPLICANT_NAME_INVALID",
-              masterName: "commonUiConfig",
-              moduleName: "patternValidation",
-              patternType: "userName",
-            },
-            patternType: "Name",
-          },
-        },
-        isMandatory: false,
-        labelChildren: "optional",
-      },
-    ],
-    dependentKey: {
-      complainantType: ["showCompanyDetails"],
-    },
-  },
-  {
-    body: [
-      {
-        key: "complainantCompanyName",
-        type: "text",
-        label: "company_Name",
-        populators: {
-          name: "complainantCompanyName",
-          error: "FIRST_LAST_NAME_MANDATORY_MESSAGE",
-          styles: {
-            minWidth: "100%",
-          },
-          validation: {
-            title: "",
-            pattern: {
-              message: "CORE_COMMON_NON_NUMERIC_TYPE_IN",
-              masterName: "commonUiConfig",
-              moduleName: "patternValidation",
-              patternType: "nonNumericString",
-            },
-            minLength: 1,
-            patternType: "Name",
-          },
-          customStyle: {
-            minWidth: "100%",
-          },
-          labelStyles: {
-            padding: "8px",
-          },
-        },
-        isMandatory: true,
-      },
-      {
-        key: "companyDetailsUpload",
-        type: "component",
-        label: "COMPANY_DOCUMENT_DETAILS",
-        component: "SelectCustomDragDrop",
-        populators: {
-          inputs: [
-            {
-              name: "document",
-              type: "DragDropComponent",
-              isOptional: "CS_IS_OPTIONAL",
-              documentHeader: "COMPANY_DOCUMENT_DETAILS",
-              isMultipleUpload: true,
-              documentHeaderStyle: {
-                textAlign: "start",
-              },
-            },
-          ],
-        },
-        isMandatory: false,
-        withoutLabel: true,
-      },
-    ],
-    head: "CS_RESPONDENT_COMPANY_DETAIL",
-    dependentKey: {
-      complainantType: ["showCompanyDetails"],
-    },
-  },
-  {
-    body: [
-      {
-        key: "addressDetails",
-        type: "component",
-        label: "PERMANENT_ADDRESS",
-        notes: fileCaseAddressMatchIdNote,
-        addUUID: true,
-        component: "SelectComponents",
-        populators: {
-          inputs: fileCaseAddressLineInputs,
-          validation: {},
-        },
-        withoutLabel: true,
-      },
-      {
-        key: "currentAddressDetails",
-        type: "component",
-        label: "CURRENT_RESIDENTIAL_ADDRESS",
-        addUUID: true,
-        component: "SelectComponents",
-        populators: {
-          inputs: fileCaseCurrentResidentialAddressInputs,
-          validation: {},
-        },
-        withoutLabel: true,
-      },
-    ],
-    head: "COMPLAINANT_ADDRESS",
-    dependentKey: {
-      complainantType: ["complainantLocation"],
-    },
-  },
-  {
-    body: [
-      {
-        key: "addressCompanyDetails",
-        type: "component",
-        addUUID: true,
-        component: "SelectComponents",
-        populators: {
-          inputs: fileCaseCompanyAddressLineInputs,
-          validation: {},
-        },
-        withoutLabel: true,
-      },
-    ],
-    head: "CS_COMPANY_LOCATION",
-    dependentKey: {
-      complainantType: ["showCompanyDetails"],
-    },
-  },
+  fileCaseComplainantTypeFormStep,
+  complainantEntityTypeFormStep,
+  complainantVerificationFormStep,
+  complainantIdFormStep,
+  fileCaseComplainantNameAgeFormStep,
+  complainantDesignationFormStep,
+  buildComplainantCompanyDetailStep(false),
+  fileCaseComplainantAddressFormStep,
+  fileCaseComplainantCompanyAddressFormStep,
   {
     body: [
       {
@@ -665,6 +349,7 @@ const complainantDetailsFormConfig = [
       marginBottom: 0,
     },
   },
+
 ];
 
 export const complaintdetailconfig = {
